requirements.txt
python-dotenv
os
shutil
hashlib
argparse